Refactor code into Stars component
