+++ client/src/components/stores/StoreIndexItem.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { withRouter } from 'react-router-dom';
+import Stars from '../ui/Stars';
 
 class StoreIndexItem extends React.Component {
   render() {
@@ -15,6 +16,7 @@ class StoreIndexItem extends React.Component {
           style={{ backgroundImage: `url(${store.image})` }}>
         </div>
         <div className="pi-name">{store.name}</div>
+        <Stars store={store} />
       </li>
     )
   }
